% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barycenter.R
\name{centroid_sphBregman_GMM}
\alias{centroid_sphBregman_GMM}
\title{Find barycenter across GMMs}
\usage{
centroid_sphBregman_GMM(stride, instanceW, supp, w, c0, options)
}
\arguments{
\item{stride}{An integer vector of the form (J_1,...,J_N) that indicates
the number of components of each GMMs.}

\item{instanceW}{A vector of 1 with length of stride}

\item{supp}{A matrix that stores the mean and covariance for all the GMMs.
More specifically, the dimension is (d+d\emph{d)x(J_1+J_2+...+J_N),
where d is the dimension of the original data.
For each column of supp, the first d rows store the mean vector,
and the last d}d rows store the vectorized covariance matrix for each component.}

\item{w}{A joined vector of weights across GMMs}

\item{c0}{GMM barycenter initialization. Usually, it can be set the same as the first GMM
or SCT}

\item{options}{list controlling pre-set parameters. \itemize{'tau'} default is 10.
\itemize{'badmm_tol'} default is 1e-4. \itemize{'badmm_max_iters'} default is 2000.
\itemize{'badmm_rho'} default is 10*mean(median(C))
where C is the total Wasserstain distance across all GMMs.}
}
\value{
Returns a list that contains MAW barycneter as well as the optimal transport
between barycenter and each GMM.
}
\description{
Find the barycenter across a set of GMMs.
The barycenter can later be used to integrate the GMMs.
}
\details{
The main steps of this procedure are outlined below. For a more detailed
description of the methodology, please see Lin L, Shi W, Ye J, Li J. (2023):
\doi{10.1111/biom.13630}
}
\examples{
\dontrun{
library(RMAW)
data("mouse_2")

# mouse_2 is a an example data file containing 8 GMMs mean and variance
m <- 15 # user should specify the number of components for MAW barycenter
stride <- t(mouse_2$stride)
mouse_2$ww <- t(mouse_2$ww)
N <- length(stride)
instanceW <- matrix(1, nrow = 1, ncol = N)
c0 <- list()
c0$supp <- mouse_2$supp[, 1:m]
c0$w <- mouse_2$ww[1:m] / sum(mouse_2$ww[1:m])
options <- list()
options$badmm_max_iters <- 1000
options$badmm_rho <- 10

result <- centroid_sphBregman_GMM(stride, instanceW, mouse_2$supp, mouse_2$ww, c0, options)

}

}
\references{
Lin L, Shi W, Ye J, Li J. (2023)
Multisource single-cell data integration by MAW barycenter for Gaussian mixture models.
Biometrics, 79, 866–877. https://doi.org/10.1111/biom.13630
}
\concept{MAW}
