% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{opt}
\alias{opt}
\title{Solve complex optimization problems,
especailly linear programs (LPs) constrained optimization.}
\usage{
opt(c, A, blc, buc, blx, bux)
}
\arguments{
\item{c:}{Objective function coefficients (vector)}

\item{A:}{Constraints matrix}

\item{blc:}{Lower bounds on constraints}

\item{buc:}{Upper bounds on constraints}

\item{blx:}{Lower bounds on variables}

\item{bux:}{Upper bounds on variables}
}
\value{
Returns a list that contains MAW distance between two GMMs as well as
the optimal transport between them.
}
\description{
Solve complex optimization problems,
especailly linear programs (LPs) constrained optimization.
}
\examples{
\dontrun{
library(RMAW)
system.file("extdata", "mouse_2.mat", package = "myfirstpackage")

# mouse_2 is a an example data file containing 8 GMMs mean and variance
stride <- mouse_2$stride
supp1 <- mouse_2$supp
result <- Mawdist(d = 2,
                  supp1 = mouse$supp[,1:stride[1]],
                  supp2 = mouse$supp[,(stride[1]+1):(stride[1]+stride[2])],
                  w1 = mouse$ww[1:stride[1]],
                  w2 = mouse$ww[(stride[1]+1):(stride[1]+stride[2])])


}

}
\concept{MAW ot}
