% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barycenter.R
\name{sqrtm_old}
\alias{sqrtm_old}
\title{Compute the square root of a matrix}
\usage{
sqrtm_old(A)
}
\arguments{
\item{A}{the matrix that is about to be computed for square root.
Usually, it should be a positive definite symmetric matrix, e.g. covariance matrix}
}
\value{
Returns the square root of the input matrix
}
\description{
Compute the square root of a matrix
}
